Introduce (memoized) term base class
This implements some of the methods shared over all memoized Node
descendants. Though RDF would only contain Named- blank-nodes and
literals, RDFlib also allows Collection and Graph types to be introduced
into a triple, so an additional layer of is introduced to overwrite some
of the core methods with more performant versions which take advantage
of the memoization (e.g. `equals()` using a single `===`).

Another major change is the introduction of `#generateString`, which
contains implementation details about how a class should be stringified.

Since there should only ever be a single instance for every unique node,
the `#toString` method is overridden to memoize the result of the first
time `#generateString` is called, saving later string concatenation
operations and garbage collection for a persistent memory allocation.