fixes #324: Fetcher.putback for contentType application/ld+json fixed
The issue concerned Fetcher.putback not passing a callback to serialize, which is required for contentType application/ld+json . The fix checks for the contentType, and if it is application/ld+json, it calls serialize with a callback, and only calls webOperation if the callback returns without an error. This seemed to be the only way to make it work without reworking serialize. serialize is sync for other contentTypes, so this check is necessary here.