Let formula, store/indexedFormula use a datafactory
The current implementation contains some methods which are like the spec
DataFactory interface but have small distinctions. E.g. `namedNode` is
called `sym`, `literal` accepts three rather than two arguments. This
commit applies the canonical methods of the rdfjs tf DataFactory
interface to the store. End-users can also overwrite the used factory
via a constructor argument.

This also replaces the use of sameTerm for the spec version equals, which
was already present in the code (sameTerm returned the result of equals).