Re-introduce workaround marking NSS as editable
As described at the following URL, the removal of this code
exposed a bug in Node Solid Server:
https://github.com/linkeddata/rdflib.js/issues/359#issuecomment-537952239

Until a fix for that bug has been deployed to the major Pod
hosters, rdflib is unable to write to those servers unless it is
running on the same origin. Thus, this commit reverts that change.
Once the fix in NSS has made its way to the major Pod hosters, this
commit can be reverted again.