Fixing tests
Reverting some of the changes done earlier - this is hacky, since we're saying one thing (typing defaultGraph as DefaultGraph) while doing one thing (creating defaultGraph as a NamedNode).

Probably needs a more thorough fix, but this was the quickest way I found, and I think it's good enough for now.